backend/requirements.txt
flask
flask-cors
selenium
beautifulsoup4
google-generativeai
python-dotenv
webdriver_manager
gunicorn 